Refactor ball spawning and paddle attachment code
Refactored the ball spawn code by moving it to player object instead which makes more sense and rewriting the way it works.
Now a spawned ball will still be attached to the paddle and to launch it you have to press the launch button.
I also changed the way paddle attachment was handled so it's less confusing and more automated.
The ball offset variable was moved to the player because it wasn't used by the ball.
Made the functions that change ball velocity not work until the ball is launched.
Changed the way the player is assigned to the ball, it's automatic now.